First slew of adjustments for arm64 port.
 * Do not override linker options set by build.kern.mk.
 * Move thread context creation to machine dependent code.
 * Fix type difference between mips and arm64 induced warnings.
 * Replace elf format definitions with those from NetBSD.
 * Provide machine independent header files with integer types information.

+++ include/sys/vm_object.h
@@ -17,7 +17,7 @@ vm_object_t *vm_object_alloc(vm_pgr_type_t type);
 void vm_object_free(vm_object_t *obj);
 bool vm_object_add_page(vm_object_t *obj, off_t offset, vm_page_t *pg);
 void vm_object_remove_page(vm_object_t *obj, vm_page_t *pg);
-void vm_object_remove_range(vm_object_t *obj, off_t offset, unsigned length);
+void vm_object_remove_range(vm_object_t *obj, off_t offset, size_t length);
 vm_page_t *vm_object_find_page(vm_object_t *obj, off_t offset);
 vm_object_t *vm_object_clone(vm_object_t *obj);
 void vm_map_object_dump(vm_object_t *obj);
